Fixed scF calculation

+++ Source/FDN.cpp
@@ -473,7 +473,7 @@ void FDN::setScatteringMatrix (MatrixType matType)
         {
 			if (FDNorder == 16)
 			{
-				std::vector<std::vector<double>> tmpMatrix(4, std::vector<double>(4, -1));
+				std::vector<std::vector<double>> tmpMatrix (4, std::vector<double>(4, -1));
 				for (int i = 0; i < 4; ++i)
 					tmpMatrix[i][i] = 1;
 
@@ -491,8 +491,10 @@ void FDN::setScatteringMatrix (MatrixType matType)
 			}
 			else
 			{
-				double scF = 2 / FDNorder;
-				std::vector<std::vector<double>> tmpMatrix(FDNorder, std::vector<double>(FDNorder, scF));
+//                double scF = 2 / FDNorder;
+                double scF = 2.0 / static_cast<double> (FDNorder);
+//                double test = 2 / FDNorder;
+				std::vector<std::vector<double>> tmpMatrix (FDNorder, std::vector<double>(FDNorder, scF));
 				for (int i = 0; i < FDNorder; ++i)
 				{
 					for (int j = 0; j < FDNorder; ++j)
